workaround for IE8 losing connection to DOM
Without this in IE8, it looks like the page hasn't loaded because you get the 'element does not exist' exception but in fact it's because something like a redirect happens and the ie.document is somehow no longer valid and throws win32OLE exceptions. This change works around that be reattaching to the window after the navigate is complete.